refactor: recover watcher
