+++ packages/mask/src/social-network-adaptor/facebook.com/collecting/posts.tsx
@@ -1,16 +1,16 @@
+import { None, Some, type Option } from 'ts-results-es'
+import { Flags } from '@masknet/flags'
 import type { SocialNetworkUI as Next } from '@masknet/types'
 import { DOMProxy, LiveSelector, MutationObserverWatcher } from '@dimensiondev/holoflows-kit'
+import { type TypedMessage, makeTypedMessageText, makeTypedMessageTuple } from '@masknet/typed-message'
 import { creator } from '../../../social-network/utils.js'
 import { isMobileFacebook } from '../utils/isMobile.js'
 import { getProfileIdentifierAtFacebook } from '../utils/getProfileIdentifier.js'
-import { type TypedMessage, makeTypedMessageText, makeTypedMessageTuple } from '@masknet/typed-message'
 import { clickSeeMore } from '../injection/PostInspector.js'
 import { facebookShared } from '../shared.js'
 import { createRefsForCreatePostContext } from '../../../social-network/utils/create-post-context.js'
 import { collectNodeText } from '../../../utils/index.js'
-import { None, Some, type Option } from 'ts-results-es'
-import { Flags } from '@masknet/flags'
-import { startWatch } from '@masknet/theme'
+import { startWatch } from '../../../utils/startWatch.js'
 
 const posts = new LiveSelector().querySelectorAll<HTMLDivElement>(
     isMobileFacebook ? '.story_body_container > div' : '[role=article]  [id]  span[dir="auto"]',
