+++ packages/mask/src/social-network-adaptor/facebook.com/injection/Composition.tsx
@@ -1,5 +1,5 @@
 import { useCallback } from 'react'
-import { makeStyles, startWatch } from '@masknet/theme'
+import { makeStyles } from '@masknet/theme'
 import { LiveSelector, MutationObserverWatcher } from '@dimensiondev/holoflows-kit'
 import { CrossIsolationMessages } from '@masknet/shared-base'
 import { attachReactTreeWithContainer } from '../../../utils/shadow-root/renderInShadowRoot.js'
@@ -7,6 +7,7 @@ import { Composition } from '../../../components/CompositionDialog/Composition.j
 import { isMobileFacebook } from '../utils/isMobile.js'
 import { PostDialogHint } from '../../../components/InjectedComponents/PostDialogHint.js'
 import { taskOpenComposeBoxFacebook, taskCloseNativeComposeBoxFacebook } from '../automation/openComposeBox.js'
+import { startWatch } from '../../../utils/startWatch.js'
 
 let composeBox: LiveSelector<Element>
