+++ packages/mask/src/social-network-adaptor/facebook.com/injection/ProfileCover.tsx
@@ -1,8 +1,8 @@
 import { MutationObserverWatcher } from '@dimensiondev/holoflows-kit'
 import { searchFacebookProfileCoverSelector } from '../utils/selector.js'
-import { attachReactTreeWithContainer } from '../../../utils/index.js'
+import { attachReactTreeWithContainer } from '../../../utils/shadow-root/renderInShadowRoot.js'
+import { startWatch } from '../../../utils/startWatch.js'
 import { ProfileCover } from '../../../components/InjectedComponents/ProfileCover.js'
-import { startWatch } from '@masknet/theme'
 
 export function injectFacebookProfileCover(signal: AbortSignal) {
     const watcher = new MutationObserverWatcher(searchFacebookProfileCoverSelector())
