+++ packages/mask/src/social-network-adaptor/facebook.com/injection/ProfileTab.tsx
@@ -1,8 +1,9 @@
 import { useEffect, useState } from 'react'
 import { debounce } from '@mui/material'
-import { makeStyles, startWatch } from '@masknet/theme'
+import { makeStyles } from '@masknet/theme'
 import { MutationObserverWatcher } from '@dimensiondev/holoflows-kit'
-import { attachReactTreeWithContainer } from '../../../utils/index.js'
+import { attachReactTreeWithContainer } from '../../../utils/shadow-root/renderInShadowRoot.js'
+import { startWatch } from '../../../utils/startWatch.js'
 import {
     profileTabSelectedSelector,
     profileTabUnselectedSelector,
