+++ packages/mask/src/social-network-adaptor/facebook.com/injection/SearchResultInspector.tsx
@@ -1,8 +1,8 @@
 import { MutationObserverWatcher } from '@dimensiondev/holoflows-kit'
 import { SearchResultInspector } from '../../../components/InjectedComponents/SearchResultInspector.js'
 import { attachReactTreeWithContainer } from '../../../utils/shadow-root/renderInShadowRoot.js'
+import { startWatch } from '../../../utils/startWatch.js'
 import { searchResultHeadingSelector } from '../utils/selector.js'
-import { startWatch } from '@masknet/theme'
 
 export function injectSearchResultInspectorAtFacebook(signal: AbortSignal) {
     const watcher = new MutationObserverWatcher(searchResultHeadingSelector())
