+++ packages/mask/src/social-network-adaptor/instagram.com/injection/Entry.tsx
@@ -1,9 +1,9 @@
+import { useState, useEffect } from 'react'
 import { Fab, styled } from '@mui/material'
 import { Create } from '@mui/icons-material'
 import { CrossIsolationMessages } from '@masknet/shared-base'
-import { Composition } from '../../../components/CompositionDialog/Composition.js'
-import { useState, useEffect } from 'react'
 import { LiveSelector, MutationObserverWatcher } from '@dimensiondev/holoflows-kit'
+import { Composition } from '../../../components/CompositionDialog/Composition.js'
 
 const Container = styled('div')`
     position: fixed;
