+++ packages/mask/src/social-network-adaptor/instagram.com/injection/NFT/NFTAvatarInInstagram.tsx
@@ -1,13 +1,14 @@
-import { type LiveSelector, MutationObserverWatcher } from '@dimensiondev/holoflows-kit'
-import { useChainContext } from '@masknet/web3-hooks-base'
-import { makeStyles, startWatch } from '@masknet/theme'
 import { max } from 'lodash-es'
 import { useEffect, useMemo, useState } from 'react'
 import { useLocation, useWindowSize } from 'react-use'
+import { type LiveSelector, MutationObserverWatcher } from '@dimensiondev/holoflows-kit'
+import { useChainContext } from '@masknet/web3-hooks-base'
+import { makeStyles } from '@masknet/theme'
+import type { AvatarMetaDB } from '@masknet/plugin-avatar'
 import { useCurrentVisitingIdentity } from '../../../../components/DataSource/useActivatedUI.js'
 import { RSS3_KEY_SNS, useNFT, useNFTAvatar, useWallet, NFTBadge, rainbowBorderKeyFrames } from '@masknet/plugin-avatar'
-import type { AvatarMetaDB } from '@masknet/plugin-avatar'
-import { attachReactTreeWithContainer } from '../../../../utils/index.js'
+import { startWatch } from '../../../../utils/startWatch.js'
+import { attachReactTreeWithContainer } from '../../../../utils/shadow-root/renderInShadowRoot.js'
 import { searchInstagramAvatarSelector } from '../../utils/selector.js'
 import { getAvatarId } from '../../utils/user.js'
