+++ packages/mask/src/social-network-adaptor/instagram.com/injection/NFT/ProfileNFTAvatar.tsx
@@ -1,9 +1,10 @@
 import { useCallback, useLayoutEffect, useState } from 'react'
 import { useLocation } from 'react-use'
 import { MutationObserverWatcher } from '@dimensiondev/holoflows-kit'
-import { makeStyles, startWatch } from '@masknet/theme'
+import { makeStyles } from '@masknet/theme'
 import { MaskMessages } from '@masknet/shared-base'
 import { useLocationChange } from '@masknet/shared-base-ui'
+import { startWatch } from '../../../../utils/startWatch.js'
 import { searchInstagramAvatarEditPageSettingDialog, searchInstagramAvatarListSelector } from '../../utils/selector.js'
 import { attachReactTreeWithContainer, useI18N } from '../../../../utils/index.js'
 import { NFTAvatarSettingDialog } from './NFTAvatarSettingDialog.js'
