+++ packages/mask/src/social-network-adaptor/instagram.com/injection/ProfileTab.tsx
@@ -2,17 +2,18 @@ import { useLayoutEffect, useMemo, useState } from 'react'
 import { useLocation } from 'react-use'
 import { MutationObserverWatcher } from '@dimensiondev/holoflows-kit'
 import { Icons } from '@masknet/icons'
-import { makeStyles, startWatch } from '@masknet/theme'
+import { makeStyles } from '@masknet/theme'
 import { MaskMessages } from '@masknet/shared-base'
 import { useMatchXS } from '@masknet/shared-base-ui'
 import { ProfileTab } from '../../../components/InjectedComponents/ProfileTab.js'
-import { attachReactTreeWithContainer } from '../../../utils/index.js'
+import { attachReactTreeWithContainer } from '../../../utils/shadow-root/renderInShadowRoot.js'
 import {
     searchProfileActiveTabSelector,
     searchProfileTabListLastChildSelector,
     searchProfileTabPageSelector,
     searchProfileTabSelector,
 } from '../utils/selector.js'
+import { startWatch } from '../../../utils/startWatch.js'
 
 export function injectProfileTabAtInstagram(signal: AbortSignal) {
     let tabInjected = false
