+++ packages/mask/src/social-network-adaptor/minds.com/injection/PostDialogHint.tsx
@@ -1,9 +1,10 @@
-import { type LiveSelector, MutationObserverWatcher } from '@dimensiondev/holoflows-kit'
-import { makeStyles, startWatch } from '@masknet/theme'
-import { CrossIsolationMessages } from '@masknet/shared-base'
 import { useCallback } from 'react'
+import { type LiveSelector, MutationObserverWatcher } from '@dimensiondev/holoflows-kit'
+import { makeStyles } from '@masknet/theme'
+import { CrossIsolationMessages } from '@masknet/shared-base'
 import { PostDialogHint } from '../../../components/InjectedComponents/PostDialogHint.js'
 import { attachReactTreeWithContainer } from '../../../utils/shadow-root/renderInShadowRoot.js'
+import { startWatch } from '../../../utils/startWatch.js'
 import { postEditorInDialogSelector, postEditorInTimelineSelector } from '../utils/selector.js'
 import { isMinds } from '../base.js'
 import { activatedSocialNetworkUI } from '../../../social-network/ui.js'
