+++ packages/mask/src/social-network-adaptor/minds.com/injection/ToolboxHint.tsx
@@ -1,8 +1,8 @@
 import { MutationObserverWatcher } from '@dimensiondev/holoflows-kit'
 import { attachReactTreeWithContainer } from '../../../utils/shadow-root/renderInShadowRoot.js'
+import { startWatch } from '../../../utils/startWatch.js'
 import { toolboxInSidebarSelector } from '../utils/selector.js'
 import { ToolboxHintAtMinds } from './ToolboxHint_UI.js'
-import { startWatch } from '@masknet/theme'
 
 export function injectToolboxHintAtMinds(signal: AbortSignal, category: 'wallet' | 'application') {
     const watcher = new MutationObserverWatcher(toolboxInSidebarSelector())
