+++ packages/mask/src/social-network-adaptor/twitter.com/injection/Avatar/index.tsx
@@ -1,11 +1,11 @@
-import { DOMProxy, MutationObserverWatcher } from '@dimensiondev/holoflows-kit'
-import { Plugin } from '@masknet/plugin-infra'
 import { noop } from 'lodash-es'
-import { Avatar } from '../../../../components/InjectedComponents/Avatar.js'
-import { attachReactTreeWithContainer } from '../../../../utils/index.js'
-import { querySelectorAll } from '../../utils/selector.js'
 import { Flags } from '@masknet/flags'
-import { startWatch } from '@masknet/theme'
+import { Plugin } from '@masknet/plugin-infra'
+import { DOMProxy, MutationObserverWatcher } from '@dimensiondev/holoflows-kit'
+import { Avatar } from '../../../../components/InjectedComponents/Avatar.js'
+import { startWatch } from '../../../../utils/startWatch.js'
+import { attachReactTreeWithContainer } from '../../../../utils/shadow-root/renderInShadowRoot.js'
+import { querySelectorAll } from '../../utils/selector.js'
 
 function getTwitterId(ele: HTMLElement) {
     const profileLink = ele.querySelector('a[role="link"]') as HTMLAnchorElement
