+++ packages/mask/src/social-network-adaptor/twitter.com/injection/Lens/injectLensOnConversation.tsx
@@ -1,10 +1,11 @@
+import { useMemo, useState } from 'react'
 import { MutationObserverWatcher } from '@dimensiondev/holoflows-kit'
 import { createInjectHooksRenderer, Plugin, useActivatedPluginsSNSAdaptor } from '@masknet/plugin-infra/content-script'
 import { EnhanceableSite, ProfileIdentifier } from '@masknet/shared-base'
-import { makeStyles, startWatch } from '@masknet/theme'
-import { useMemo, useState } from 'react'
-import { attachReactTreeWithContainer } from '../../../../utils/index.js'
+import { makeStyles } from '@masknet/theme'
+import { attachReactTreeWithContainer } from '../../../../utils/shadow-root/renderInShadowRoot.js'
 import { querySelectorAll } from '../../utils/selector.js'
+import { startWatch } from '../../../../utils/startWatch.js'
 
 const selector = () => {
     return querySelectorAll<HTMLElement>('[data-testid=conversation] div:not([tabindex]) div[dir] + div[dir]')
