+++ packages/mask/src/social-network-adaptor/twitter.com/injection/PluginSettingsDialog.tsx
@@ -1,8 +1,8 @@
 import { MutationObserverWatcher } from '@dimensiondev/holoflows-kit'
 import { PluginSettingsDialog } from '../../../components/InjectedComponents/PluginSettingsDialog.js'
-import { attachReactTreeWithContainer } from '../../../utils/index.js'
+import { attachReactTreeWithContainer } from '../../../utils/shadow-root/renderInShadowRoot.js'
+import { startWatch } from '../../../utils/startWatch.js'
 import { rootSelector } from '../utils/selector.js'
-import { startWatch } from '@masknet/theme'
 
 /**
  * @deprecated unused
