+++ packages/mask/src/social-network-adaptor/twitter.com/injection/PostDialog.tsx
@@ -2,7 +2,7 @@ import { MutationObserverWatcher, type LiveSelector } from '@dimensiondev/holofl
 import { attachReactTreeWithContainer } from '../../../utils/shadow-root/renderInShadowRoot.js'
 import { Composition } from '../../../components/CompositionDialog/Composition.js'
 import { postEditorContentInPopupSelector, rootSelector } from '../utils/selector.js'
-import { startWatch, type WatchOptions } from '@masknet/theme'
+import { startWatch, type WatchOptions } from '../../../utils/startWatch.js'
 
 function renderPostDialogTo<T>(reason: 'timeline' | 'popup', ls: LiveSelector<T, true>, options: WatchOptions) {
     const watcher = new MutationObserverWatcher(ls)
