+++ packages/mask/src/social-network-adaptor/twitter.com/injection/SwitchLogo.tsx
@@ -1,8 +1,8 @@
+import { SwitchLogoIconButton } from '@masknet/plugin-switch-logo'
 import { MutationObserverWatcher, type LiveSelector } from '@dimensiondev/holoflows-kit'
 import { querySelector } from '../utils/selector.js'
-import { startWatch } from '@masknet/theme'
-import { attachReactTreeWithContainer } from '../../../utils/index.js'
-import { SwitchLogoIconButton } from '@masknet/plugin-switch-logo'
+import { startWatch } from '../../../utils/startWatch.js'
+import { attachReactTreeWithContainer } from '../../../utils/shadow-root/renderInShadowRoot.js'
 
 const logoSelector: () => LiveSelector<HTMLElement, true> = () => {
     return querySelector<HTMLElement>('h1[role="heading"] a > div > svg').closest(1)
