+++ packages/mask/src/social-network-adaptor/twitter.com/injection/Tips/FollowTipsButton.tsx
@@ -2,11 +2,12 @@ import { memo, useMemo, useState } from 'react'
 import { useAsync } from 'react-use'
 import { noop } from 'lodash-es'
 import { Flags } from '@masknet/flags'
-import { makeStyles, startWatch } from '@masknet/theme'
+import { makeStyles } from '@masknet/theme'
 import { DOMProxy, MutationObserverWatcher } from '@dimensiondev/holoflows-kit'
 import { createInjectHooksRenderer, Plugin, useActivatedPluginsSNSAdaptor } from '@masknet/plugin-infra/content-script'
 import { useThemeSettings } from '../../../../components/DataSource/useActivatedUI.js'
-import { attachReactTreeWithContainer } from '../../../../utils/index.js'
+import { attachReactTreeWithContainer } from '../../../../utils/shadow-root/renderInShadowRoot.js'
+import { startWatch } from '../../../../utils/startWatch.js'
 import { TipButtonStyle } from '../../constant.js'
 import { normalFollowButtonSelector as selector } from '../../utils/selector.js'
 import { getUserIdentity } from '../../utils/user.js'
